+++ src/Git_Updater/Messages.php
@@ -236,30 +236,4 @@ class Messages {
 		}
 	}
 
-	/**
-	 * Git Updater PRO upsell notice.
-	 *
-	 * @return void
-	 */
-	public function show_upsell() {
-		if ( $this->is_pro_running() ) {
-			return;
-		}
-		?>
-		<div class="notice-info notice">
-			<p>
-				<?php esc_html_e( 'Git Updater PRO', 'git-updater' ); ?>
-				<br>
-				<?php
-				printf(
-					/* translators: %1: opening href tag, %2: closing href tag */
-					esc_html__( 'Unlock PRO features like remote installation, branch switching, REST API, WP-CLI, and more. Information at %1$sgit-updater.com%2$s.', 'git-updater' ),
-					'<a href="https://git-updater.com">',
-					'</a>'
-				);
-				?>
-			</p>
-		</div>
-		<?php
-	}
 }